tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/size.attributes.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: size.attributes.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.default</h1>
<p class="desc">Default width/height when attributes are missing</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" ><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.default.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Default width/height when attributes are missing");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 300, "canvas.width", "300");
_assertSame(canvas.height, 150, "canvas.height", "150");
_assert(!canvas.hasAttribute('width'), "!canvas.hasAttribute('width')");
_assert(!canvas.hasAttribute('height'), "!canvas.hasAttribute('height')");


});
</script>
